Move __vars__ content into other modules
The __vars__ module contains dundered variables that are either (a)
module-level metadata like '__version__' or (b) static API endpoint
URLs.

Condense the codebase by moving the metadata into the package __init__
and the API endpoints into the api_urls module.
